fix: move command list
